refactor(publish): one-way layering — repositories never import the publish layer
The publish pipeline was architecturally inverted: server/repositories/
publish.ts orchestrated the entire full publish (runtime builds, rendering,
Layer A bake, slot swap, cache bump) from inside the repository layer, and
repositories/data/{publish,tables}.ts pulled renderers and artefact writers
in the same backwards direction — producing 7 of the repo's circular-
dependency cycles plus a lazy `await import()` band-aid in tables.ts.

Repositories are now data access only; orchestration lives in server/publish/:

- server/publish/publishSite.ts (new) — publishDraftSite orchestrator.
  repositories/publish.ts keeps the snapshot getters, the publish-status
  check, getDraftSiteDocument (deduped draft assembly), and the short
  persistSitePublish transaction.
- server/publish/publishRow.ts (new) — publishDataRow + removeDataRowArtefact.
  repositories/data/publish.ts keeps persistDataRowPublish (the transaction)
  and the route lookups.
- server/publish/templateSeeding.ts (moved) — seeding publishes a page row,
  so it is publish-layer orchestration. createDataTable no longer auto-seeds
  via lazy import; every table-creation entry point (admin handler, plugin
  content API) calls ensureDefaultEntryTemplate explicitly, and the CMS
  bundle import runs the idempotent backfill after its transaction commits.
  Post-import backfill also fixes the replace-strategy duplicating a default
  template the bundle already carried, and heals bundles imported without one.

All publish-pipeline circular dependencies are gone (madge: 27 → 20 repo-
wide; the remaining ones are the unrelated module index↔editor lazy pattern
and small core pairs). Architecture gates (publish-bumps-cache-version,
dataRowVersionAllocator) updated to the new file map; docs updated in step.
