+++ docs/features/content-storage.md
@@ -165,8 +165,8 @@ This is the single source of truth for slug derivation used by all admin write p
 | `server/repositories/data/rows/import.ts`        | Bundle-import upserts (id-preserving): `upsertDataRow`, `insertDataRowIfAbsent`, `replaceDataRow` |
 | `server/repositories/data/rows/mapper.ts`        | Internal: hydrated SELECT builder + `DataRowRow → DataRow` mapper (not part of the public barrel) |
 | `server/repositories/data/rows/index.ts`         | Barrel for the `rows/` directory                                     |
-| `server/repositories/data/publish.ts`            | Publish / unpublish / schedule a row; write `data_row_versions`       |
-| `server/repositories/data/templateSeeding.ts`    | Seed default entry template for new postType tables                   |
+| `server/repositories/data/publish.ts`            | Publish persistence (`persistDataRowPublish` writes `data_row_versions`) + public-route lookups; the orchestration (lock, artefacts, cache bump) is `server/publish/publishRow.ts` |
+| `server/publish/templateSeeding.ts`              | Seed default entry template for new postType tables (publish layer — it publishes a page row) |
 | `server/repositories/data/shared.ts`             | Shared helpers: `userRefAt` (typed accessor per prefix — unknown prefix is a compile error), `userRefColumns` / `userRefJoin` (SQL fragment builders — the single source for the four `<prefix>_*` user-ref join columns and LEFT JOIN clauses, spliced verbatim by both `rows/mapper.ts` and `publish.ts`), `UserJoinColumns` (interface for all four `<prefix>_*` column groups — always present via LEFT JOIN, `null` when no user matched) |
 | `server/repositories/data/index.ts`              | Barrel for the whole `data/` directory                               |
 
@@ -348,9 +348,10 @@ Events are emitted from `server/publish/contentEvents.ts`, which also exports `a
   - `src/core/data/fields.ts` — field normalization, built-in field detection
   - `src/core/data/pageFromRow.ts` — Page ↔ row
   - `src/core/data/componentFromRow.ts` — VC ↔ row
-  - `server/repositories/data/` — `tables.ts`, `rows/` (split by responsibility), `publish.ts`, `templateSeeding.ts`, `shared.ts`
+  - `server/repositories/data/` — `tables.ts`, `rows/` (split by responsibility), `publish.ts`, `shared.ts`
   - `server/handlers/cms/data/` — generic data endpoints
   - `server/handlers/cms/pages.ts`, `components.ts`, `layouts.ts` — typed endpoints for the system tables
+  - `server/publish/publishRow.ts` — per-row publish orchestration; `server/publish/templateSeeding.ts` — default entry-template seeding
   - `server/publish/publishScheduler.ts` — scheduled-publish tick
 - Gate tests:
   - `src/__tests__/architecture/data-tables-system-flag.test.ts`
