+++ scripts/bench/benches/publish.ts
@@ -46,7 +46,8 @@ async function loadServer() {
   const { runMigrations } = await import('../../../server/db/runMigrations')
   const { sqliteMigrations } = await import('../../../server/db/migrations-sqlite')
   const { saveDraftSite } = await import('../../../server/repositories/site')
-  const { publishDraftSite, getDraftPublishStatus } = await import('../../../server/repositories/publish')
+  const { getDraftPublishStatus } = await import('../../../server/repositories/publish')
+  const { publishDraftSite } = await import('../../../server/publish/publishSite')
   const { createDataRow, listDataRows, listDataRowIdSlugs, updateDataRowDraftCells } = await import('../../../server/repositories/data')
   const { getPublishedDataRowByRoute } = await import('../../../server/repositories/data/publish')
   const { getSetupStatus } = await import('../../../server/repositories/setup')
