+++ server/handlers/cms/publish.ts
@@ -21,7 +21,8 @@
 import type { DbClient } from '../../db/client'
 import { requireCapability, requireStepUp } from '../../auth/authz'
 import { createAuditEvent } from '../../repositories/audit'
-import { getDraftPublishStatus, publishDraftSite } from '../../repositories/publish'
+import { getDraftPublishStatus } from '../../repositories/publish'
+import { publishDraftSite } from '../../publish/publishSite'
 import { jsonResponse, methodNotAllowed } from '../../http'
 import type { CmsHandlerOptions } from './shared'
 import { requestAuditContext } from './shared'
