+++ server/index.ts
@@ -1,7 +1,7 @@
 import { createDbClient } from './db'
 import { runMigrations } from './db/runMigrations'
 import { syncSystemRoles } from './repositories/roles'
-import { backfillDefaultEntryTemplates } from './repositories/data'
+import { backfillDefaultEntryTemplates } from './publish/templateSeeding'
 import { readServerConfig } from './config'
 import { DEV_ORIGIN_ALLOWLIST, configurePublicOrigins, configureTrustedProxyCidrs, stampSocketIp } from './auth/security'
 import { startConversationPurgeTick } from './ai/boot'
@@ -22,9 +22,10 @@ await runMigrations(db, migrations)
 // capabilities are added in code. See `syncSystemRoles` for the policy.
 await syncSystemRoles(db)
 // Every postType data table needs a default entry template so the public
-// route `/<route-base>/<row-slug>` resolves. `createDataTable` seeds one
-// at creation time; this backfill covers the baseline-seeded `posts` table
-// and any postType row that pre-dates the seeding feature.
+// route `/<route-base>/<row-slug>` resolves. Table-creation entry points
+// seed one right after `createDataTable`; this backfill covers the
+// baseline-seeded `posts` table and any postType row that pre-dates the
+// seeding feature.
 await backfillDefaultEntryTemplates(db)
 // Wire the built-in local-disk media adapter BEFORE plugins activate —
 // plugin adapters register through the same registry but local-disk is
