+++ server/repositories/data/tables.ts
@@ -218,23 +218,12 @@ export async function createDataTable(
               primary_field_id, fields_json, system,
               created_by_user_id, updated_by_user_id, created_at, updated_at
   `
-  const table = mapTable(rows[0])
-
-  // Every postType table needs a public-facing entry template — otherwise
-  // `/<route-base>/<row-slug>` 404s on first publish. We seed a minimal
-  // default here so the public route works immediately; site owners then
-  // customise the template in the editor. The seed is idempotent — re-runs
-  // (e.g. via the boot backfill) are no-ops.
-  //
-  // Lazy import avoids a circular dependency: `templateSeeding` calls
-  // `createDataRow` + `publishDataRow` which live in sibling files that
-  // depend on this table module.
-  if (table.kind === 'postType') {
-    const { ensureDefaultEntryTemplate } = await import('./templateSeeding')
-    await ensureDefaultEntryTemplate(db, table, input.createdByUserId ?? null)
-  }
-
-  return table
+  // NOTE: table creation does NOT seed the default entry template — seeding
+  // publishes a page row, which is publish-layer orchestration. Every
+  // table-creation entry point calls `ensureDefaultEntryTemplate`
+  // (`server/publish/templateSeeding.ts`) right after this returns, and the
+  // boot/import backfill covers every other path.
+  return mapTable(rows[0])
 }
 
 export async function updateDataTable(
